requirements.txt
social-auth-app-django==5.0.0
requests==2.28.2
python-dotenv==1.0.0